refactor: send fcm token only once
