+++ src/hooks/pushAlarm/useGetFCMTokenFromApp.ts
@@ -1,10 +1,15 @@
+import { useSetRecoilState } from 'recoil';
+
 import useListeningAppMessage from '../bridge/useListeningAppMessage';
 
+import fcmTokenState from '@/store/push-notification/fcmToken';
+
 const useGetFCMTokenFromApp = () => {
+  const setFcmToken = useSetRecoilState(fcmTokenState);
   useListeningAppMessage({
     targetType: 'FCM_TOKEN',
     handler: ({ data }) => {
-      localStorage.setItem('fcm_token', JSON.stringify(data));
+      setFcmToken(data as string);
     },
   });
 };
