+++ src/pages/_app.page.tsx
@@ -10,9 +10,9 @@ import { domMax, LazyMotion } from 'framer-motion';
 import { RecoilRoot } from 'recoil';
 
 import ToastWrapper from '@/components/portal/ToastWrapper';
+import FcmTokenHandler from '@/components/push-notification/FcmTokenHandler';
 import RouteGuard from '@/components/route-guard/RouteGuard';
 import useTrackPageView from '@/hooks/analytics/useTrackPageView';
-import useGetFCMTokenFromApp from '@/hooks/pushAlarm/useGetFCMTokenFromApp';
 import GlobalStyles from '@/styles/GlobalStyles';
 import lightTheme from '@/styles/theme';
 
@@ -40,7 +40,6 @@ const MyApp = ({ Component: AppComponent, pageProps }: AppPropsWithLayout) => {
   );
 
   useTrackPageView();
-  useGetFCMTokenFromApp();
 
   return (
     <QueryClientProvider client={queryClient}>
@@ -48,12 +47,14 @@ const MyApp = ({ Component: AppComponent, pageProps }: AppPropsWithLayout) => {
         <RecoilRoot>
           <ThemeProvider theme={lightTheme}>
             <LazyMotion features={domMax}>
-              <DefaultLayout>
-                <Head />
-                <GlobalStyles />
-                <RouteGuard>{getLayout(<AppComponent {...pageProps} />)}</RouteGuard>
-                <ToastWrapper />
-              </DefaultLayout>
+              <FcmTokenHandler>
+                <DefaultLayout>
+                  <Head />
+                  <GlobalStyles />
+                  <RouteGuard>{getLayout(<AppComponent {...pageProps} />)}</RouteGuard>
+                  <ToastWrapper />
+                </DefaultLayout>
+              </FcmTokenHandler>
             </LazyMotion>
           </ThemeProvider>
         </RecoilRoot>
