+++ src/store/push-notification/fcmToken.ts
@@ -1,11 +1,8 @@
 import { atom } from 'recoil';
 
-import localStorageEffect from './effect';
-
 const fcmTokenState = atom<string | null>({
   key: 'fcmToken',
   default: null,
-  effects: [localStorageEffect<string | null>('fcm_token')],
 });
 
 export default fcmTokenState;
